Refactoring: moving debug-related var and func to common file
